create new class, copy setup, move methods
