+++ tests/Backend/CommonPart1/TableWithConfigurationLoadTest.php
@@ -6,19 +6,14 @@ use Keboola\Csv\CsvFile;
 use Keboola\StorageApi\Client;
 use Keboola\StorageApi\ClientException;
 use Keboola\StorageApi\Components;
-use Keboola\StorageApi\Options\Components\Configuration;
 use Keboola\StorageApi\Options\FileUploadOptions;
 use Keboola\Test\Backend\TableWithConfigurationUtils;
-use Keboola\Test\Backend\WorkspaceConnectionTrait;
-use Keboola\Test\Backend\Workspaces\Backend\WorkspaceBackendFactory;
-use Keboola\Test\Backend\Workspaces\ParallelWorkspacesTestCase;
 use Keboola\Test\ClientProvider\ClientProvider;
 use Keboola\Test\StorageApiTestCase;
 
-class TableWithConfigurationLoadTest extends ParallelWorkspacesTestCase
+class TableWithConfigurationLoadTest extends StorageApiTestCase
 {
     use TableWithConfigurationUtils;
-    use WorkspaceConnectionTrait;
 
     public const DEFAULT_CONFIGURATION_MIGRATIONS = [
         [
@@ -89,44 +84,6 @@ class TableWithConfigurationLoadTest extends ParallelWorkspacesTestCase
         $this->_client->writeTableAsyncDirect($tableId, [
             'dataFileId' => $fileId,
         ] + $writeTableOptions);
-    }
-
-    private function loadTableFromWorkspace(
-        string $tableId,
-        array $writeTableOptions = []
-    ): void {
-        // create workspace and source table in workspace
-        $workspace = $this->initTestWorkspace();
-
-        $sourceTableName = 'languages';
-        $backend = WorkspaceBackendFactory::createWorkspaceBackend($workspace);
-        $backend->dropTableIfExists($sourceTableName);
-
-        $connection = $workspace['connection'];
-
-        $db = $this->getDbConnectionSynapse($connection);
-
-        $quotedSourceTableId = $db->getDatabasePlatform()->quoteIdentifier(sprintf(
-            '%s.%s',
-            $connection['schema'],
-            $sourceTableName,
-        ));
-
-        // data from `../../_data/languages.csv`
-        $db->executeQuery("CREATE TABLE $quotedSourceTableId (
-			[Id] INTEGER,
-			[NAME] VARCHAR(100)
-		);");
-        $db->executeQuery("INSERT INTO $quotedSourceTableId ([id], [NAME]) SELECT 0, '- unchecked -'");
-        $db->executeQuery("INSERT INTO $quotedSourceTableId ([id], [NAME]) SELECT 26, 'czech'");
-        $db->executeQuery("INSERT INTO $quotedSourceTableId ([id], [NAME]) SELECT 1, 'english'");
-        $db->executeQuery("INSERT INTO $quotedSourceTableId ([id], [NAME]) SELECT 11, 'finnish'");
-        $db->executeQuery("INSERT INTO $quotedSourceTableId ([id], [NAME]) SELECT 24, 'french'");
-
-        $this->_client->writeTableAsyncDirect($tableId, [
-                'dataWorkspaceId' => $workspace['id'],
-                'dataObject' => $sourceTableName,
-            ] + $writeTableOptions);
     }
 
     public function testLoadFromFileToTable(): void
@@ -570,95 +527,6 @@ JSON;
         // check events
         $events = $this->listEventsFilteredByName($this->client, 'storage.tableWithConfigurationImportQuery', $tableId, 50);
         $this->assertCount(9, $events);
-
-        $events = $this->listEventsFilteredByName($this->client, 'storage.tableImportDone', $tableId, 10);
-        $this->assertCount(1, $events);
-    }
-
-    public function testLoadFromWorkspaceToTable(): void
-    {
-        $tableName = 'custom-table-3';
-
-        $json = /** @lang JSON */
-            <<<JSON
-{
-  "action": "generate",
-  "backend": "synapse",
-  "operation": "importFull",
-  "source": "table",
-  "columns": [
-    "id",
-    "NAME"
-  ],
-  "primaryKeys": [
-    "id"
-  ],
-  "output": {
-    "queries": [
-      {
-        "sql": "CREATE TABLE {{ id(destSchemaName) }}.{{ id(stageTableName) }} ([id] NVARCHAR(4000), [NAME] NVARCHAR(4000)) WITH (DISTRIBUTION = ROUND_ROBIN,CLUSTERED COLUMNSTORE INDEX)",
-        "description": ""
-      },
-      {
-        "sql": "INSERT INTO {{ id(destSchemaName) }}.{{ id(stageTableName) }} ([id], [NAME]) SELECT [id], [NAME] FROM {{ id(sourceSchemaName) }}.{{ id(sourceTableName) }}",
-        "description": ""
-      },
-      {
-        "sql": "CREATE TABLE {{ id(destSchemaName) }}.{{ id(destTableName ~ rand ~ '_tmp') }} WITH (DISTRIBUTION=ROUND_ROBIN,CLUSTERED COLUMNSTORE INDEX) AS SELECT a.[id],a.[NAME] FROM (SELECT COALESCE([id], '') AS [id],COALESCE([NAME], '') AS [NAME], ROW_NUMBER() OVER (PARTITION BY [id] ORDER BY [id]) AS \"_row_number_\" FROM {{ id(destSchemaName) }}.{{ id(stageTableName) }}) AS a WHERE a.\"_row_number_\" = 1",
-        "description": ""
-      },
-      {
-        "sql": "RENAME OBJECT {{ id(destSchemaName) }}.{{ id(destTableName) }} TO {{ id(destTableName ~ rand ~ '_tmp_rename') }}",
-        "description": ""
-      },
-      {
-        "sql": "RENAME OBJECT {{ id(destSchemaName) }}.{{ id(destTableName ~ rand ~ '_tmp') }} TO {{ id(destTableName) }}",
-        "description": ""
-      },
-      {
-        "sql": "DROP TABLE {{ id(destSchemaName) }}.{{ id(destTableName ~ rand ~ '_tmp_rename') }}",
-        "description": ""
-      },
-      {
-        "sql": "IF OBJECT_ID (N'{{ id(destSchemaName) }}.{{ id(destTableName ~ rand ~ '_tmp') }}', N'U') IS NOT NULL DROP TABLE {{ id(destSchemaName) }}.{{ id(destTableName ~ rand ~ '_tmp') }}",
-        "description": ""
-      },
-      {
-        "sql": "IF OBJECT_ID (N'{{ id(destSchemaName) }}.{{ id(destTableName ~ rand ~ '_tmp_rename') }}', N'U') IS NOT NULL DROP TABLE {{ id(destSchemaName) }}.{{ id(destTableName ~ rand ~ '_tmp_rename') }}",
-        "description": ""
-      }
-    ]
-  }
-}
-JSON;
-
-        [$tableId,] = $this->createTableWithConfiguration($json, $tableName, 'outputMappingFullLoad');
-
-        $this->loadTableFromWorkspace(
-            $tableId,
-        );
-
-        $table = $this->_client->getTable($tableId);
-
-        $this->assertEquals(['id', 'NAME'], $table['columns']);
-        $this->assertSame(5, $table['rowsCount']);
-        $this->assertTableColumnMetadata([
-            'id' => [
-                'KBC.datatype.type' => 'INT',
-                'KBC.datatype.nullable' => '1',
-                'KBC.datatype.basetype' => 'INTEGER',
-            ],
-            'NAME' => [
-                'KBC.datatype.type' => 'VARCHAR',
-                'KBC.datatype.nullable' => '1',
-                'KBC.datatype.basetype' => 'STRING',
-                'KBC.datatype.length' => '100',
-            ],
-        ], $table);
-
-        // check events
-        $events = $this->listEventsFilteredByName($this->client, 'storage.tableWithConfigurationImportQuery', $tableId, 50);
-        $this->assertCount(8, $events);
 
         $events = $this->listEventsFilteredByName($this->client, 'storage.tableImportDone', $tableId, 10);
         $this->assertCount(1, $events);
